backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pillow-heif
opencv-python
numpy
reportlab
openai
google-generativeai